+++ PouEngine/include/PouEngine/scene/SceneObject.h
@@ -25,13 +25,13 @@ class SceneObject : public NotificationListener
         bool isAnEntity();
         bool isAShadowCaster();
 
-        void setLocalTime(float time);
-        virtual void setSyncAndLocalTime(float time);
-        void setLastUpdateTime(float time, bool force = false);
-        float getLastUpdateTime();
-        float getLastNodeUpdateTime();
+        void setLocalTime(uint32_t time);
+        virtual void setSyncAndLocalTime(uint32_t time);
+        void setLastUpdateTime(uint32_t time, bool force = false);
+        uint32_t getLastUpdateTime();
+        uint32_t getLastNodeUpdateTime();
 
-        virtual void update(const Time &elapsedTime, float localTime = -1);
+        virtual void update(const Time &elapsedTime, uint32_t localTime = -1);
         virtual void notify(NotificationSender* , NotificationType,
                             size_t dataSize = 0, char* data = nullptr) override;
 
@@ -44,10 +44,10 @@ class SceneObject : public NotificationListener
         bool m_isAnEntity;
         bool m_isAShadowCaster;
 
-        float m_curLocalTime;
-        float m_lastSyncTime;
-        float m_lastUpdateTime;
-        float m_lastNodeUpdateTime;
+        uint32_t m_curLocalTime;
+        uint32_t m_lastSyncTime;
+        uint32_t m_lastUpdateTime;
+        uint32_t m_lastNodeUpdateTime;
 
     private:
 };
